# c_openapi.yaml
openapi: 3.1.0
info:
  title: 住民記録システム Web版 API
  version: 0.1.0
  description: |
    住民記録システム標準仕様書 第6.1版（令和8年3月25日版 / 自治体システム等標準化検討会）に準拠した
    REST API の設計仕様（ドラフト）。ChromeOS Flex を含む Chromium 系ブラウザからの利用を想定。

    - すべての永続化操作は監査ログに記録される。
    - すべての応答は UTF-8 / application/json。日付は ISO 8601。
    - 認可は OAuth2 / OIDC のスコープ + 属性ベース（ロール + 項目別マスク）。
servers:
  - url: https://{tenant}.juki.example.go.jp/api/v1
    variables:
      tenant: { default: city01 }
...
security:
  - oidc: [residents.read, residents.write, transactions.write, certificates.issue, admin]
paths:
...
  /restrictions:
    post:
      tags: [restrictions]
      summary: 抑止登録（DV等支援措置）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Restriction' } } } }
      responses: { '201': { description: Created } }
  /restrictions/{id}:
    delete:
      tags: [restrictions]
      summary: 抑止解除（解除権限ロールのみ）
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '204': { description: OK } }

  # ===== 統計 / EUC =====
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://idp.example.go.jp/.well-known/openid-configuration
...
  schemas:
...
    Restriction:
      type: object
      properties:
        id:          { type: string, readOnly: true }
        residentId:  { type: string }
        category:    { type: string, enum: [DV, STALKER, CHILD_ABUSE, OTHER] }
        startDate:   { type: string, format: date }
        endDate:     { type: string, format: date }
        scope:       { type: string, enum: [SELF, HOUSEHOLD] }
        releaseRole: { type: string }
        note:        { type: string }
...